cleaned up code, put input validation functions in their own module, git rid of the error in manifest.json
